Fix issue #3335 with renamed CSV files
Change External Data mechanism for importing CSV files for search()/pulldata() functionalty so that it no longer has to rename a file to track file changes. The CSV file timestamp is now captured in a metadata table stored in the SQLite database holding the imported file. Updates to the file will cause the database to be rebuilt the next time the form is loaded. If the timestamp is unchanged, we skip importing the data.

Now that the file is no longer being renamed, other ODK functionality that relies on CSV files can operate as expected without having their files disappear from under them. However, the each CSV file will still be imported into a SQLite database by the External Data functionality, since it is not implemented narrowly enough to understand which CSV files are actually referenced by the form via search()/pulldata().